requirements.txt
asgiref==3.4.1
Django==3.2
django-ckeditor==6.2.0
django-js-asset==1.2.2
mysqlclient==2.1.0
psycopg2==2.9.3
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
